requirements.txt
streamlit==1.38.0
pandas==2.2.3
python-dotenv==1.0.1